WIP: ipq806x: fix sfp aka sgmii on gmac1 and gmac2 and gmac3 in general
This is an attempt to fix dwmac operation for SFP (e.g. using SGMII) on
gmac1 and dwmac operation on gmac2 and gmac3 in general taken from [1].
However, while with this the driver comes up without errors and SFP on
gmac1 actually does work no data seems to be flowing on gmac2 and gmac3
as of yet.

[1] https://forum.openwrt.org/t/support-for-mikrotik-rb3011uias-rm/4064/257

Signed-off-by: Marcel Ziswiler <marcel@ziswiler.com>