ipq806x: add ipq-aux-loader
This adds ipq-aux-loader from Sergey Sergeev <adron@yapic.net>
It is a small loader for legacy or LZMA compressed FIT images.
This is needed to add support for Mikrotik RouterBoards as they expect a
specific ELF image.
This loader enables to use LZMA compressed FIT image which will
decompress the kernel image and boot it.

This is similar to what Robert Marko <robimarko@gmail.com> has done for
ipq40x.

Signed-off-by: Sergey Sergeev <adron@yapic.net>